Add service-layer validations, refactor service and repository layers, and update database data

+++ WalletWise.Application/Services/GenericService.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using AutoMapper;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -10,79 +11,102 @@ using WalletWise.Domain.Interfaces;
 
 namespace WalletWise.Application.Services
 {
-    public class GenericService<T> : IGenericService<T> where T : class
+    public class GenericService<T, ResDto, CreatDto, UpdaDto> : IGenericService<ResDto, CreatDto, UpdaDto> 
+        where T : class 
+        where ResDto : class
+        where CreatDto : class
+        where UpdaDto : class
     {
         private readonly IGenericRepository<T> _repository;
-        private readonly ILogger<T> _logger;
+        protected readonly ILogger<T> _logger;
+        protected readonly IMapper _mapper;
 
-        public GenericService(IGenericRepository<T> repository, ILogger<T> logger)
+        public GenericService(IGenericRepository<T> repository, ILogger<T> logger, IMapper mapper)
         {
             _repository = repository;
             _logger = logger;
+            _mapper = mapper;
         }
 
-        public virtual async Task<Result<T?>> GetByIdAsync(int id)
+        public virtual async Task<Result<ResDto?>> GetByIdAsync(int id)
         {
             try
             {
-                var entity = await _repository.GetByIdAsync(id);
+                T? entity = await _repository.GetByIdAsync(id);
 
                 if (entity == null)
                 {
-                    return Result<T?>.Failure($"La entidad con el Id {id} no pudo ser encontrada");
+                    return Result<ResDto?>.Failure($"La entidad con el Id {id} no pudo ser encontrada");
 
                 }
-                return Result<T?>.Success(entity);
+
+                return Result<ResDto?>.Success(_mapper.Map<ResDto>(entity));
 
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex + " Ha ocurrido un error al obtener al entidad con el id " + typeof(T).Name);
-                return Result<T?>.Failure("No se ha podido obtener la entidad");
+                _logger.LogError(ex, "Ha ocurrido un error al obtener la entidad {T} con id {Id}", typeof(T).Name, id);
+                return Result<ResDto?>.Failure("No se ha podido obtener la entidad");
             }
 
         }
 
-        public virtual async Task<Result<IEnumerable<T>>> GetAllAsync()
+        public virtual async Task<Result<IEnumerable<ResDto>>> GetAllAsync()
         {
             try
             {
-                return Result<IEnumerable<T>>.Success(await _repository.GetAllAsync());
+                IEnumerable<T> values = await _repository.GetAllAsync();
+
+                var a = _mapper.Map<IEnumerable<ResDto>>(values);
+
+                return Result<IEnumerable<ResDto>>.Success(a);
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex + "A ocurrido un error al obtener todas entidades");
-                return Result<IEnumerable<T>>.Failure("No se a podido listar todas las entidades");
+                _logger.LogError(ex, "A ocurrido un error al obtener todas entidades");
+                return Result<IEnumerable<ResDto>>.Failure("No se a podido listar todas las entidades");
             }
         }
 
-        public virtual async Task<Result<T>> AddAsync(T entity)
+        public virtual async Task<Result<ResDto>> AddAsync(CreatDto DtoRequest)
         {
             try
             {
-                return Result<T>.Success(
-                    await _repository.AddAsync(entity));
+                T entity = _mapper.Map<T>(DtoRequest);
+
+                T values = await _repository.AddAsync(entity);
+
+                return Result<ResDto>.Success(_mapper.Map<ResDto>(values));
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex + " A ocurrido un error inesperado al crear una entidad ", typeof(T).Name);
-                return Result<T>.Failure("No se a podido crear la entidad");
+                _logger.LogError(ex, "A ocurrido un error inesperado al crear una entidad {T}", typeof(T).Name);
+                return Result<ResDto>.Failure("No se a podido crear la entidad");
             }
         }
 
 
-        public virtual async Task<Result<T>> UpdateAsync(T entity)
+        public virtual async Task<Result<ResDto>> UpdateAsync(int id, UpdaDto dtoRequest)
         {
             try
             {
+                T? exist = await _repository.GetByIdAsync(id);
+
+                if (exist == null)
+                {
+                    return Result<ResDto>.Failure($"La entidad con el Id {id} no existe");
+                }
+
+                T entity = _mapper.Map<T>(dtoRequest);
+
                 await _repository.UpdateAsync(entity);
 
-                return Result<T>.Success(entity);
+                return Result<ResDto>.Success(_mapper.Map<ResDto>(entity));
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex + " Fallo al actualizar la entidad ", typeof(T).Name);
-                return Result<T>.Failure("No se a podido actualizar la entidad");
+                _logger.LogError(ex, "Fallo al actualizar la entidad {T}", typeof(T).Name);
+                return Result<ResDto>.Failure("No se a podido actualizar la entidad");
             }
         }
 
@@ -90,13 +114,19 @@ namespace WalletWise.Application.Services
         {
             try
             {
+                var exist = await _repository.GetByIdAsync(id);
+
+                if (exist == null)
+                {
+                    return Result<bool>.Failure($"La entidad con el Id {id} no pudo ser encontrada");
+                }
 
                 await _repository.RemoveAsync(id);
                 return Result<bool>.Success(true);
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, " A ocurrido un fallo al borrar la entidad " + id);
+                _logger.LogError(ex, "A ocurrido un fallo al borrar la entidad con el id {Id}", id);
                 return Result<bool>.Failure("No se ha podido eliminar la entidad");
             }
         }
